refactor: replace block Modal forms with Campaigns version
