+++ src/Campaigns/Blocks/CampaignForm/resources/components/BlockPreview.tsx
@@ -1,7 +1,7 @@
 import { useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import IframeResizer from 'iframe-resizer-react';
-import ModalForm from './ModalForm';
+import ModalForm from '../../../shared/components/ModalForm';
 import ServerSideRender from '@wordpress/server-side-render';
 
 import '../../../shared/components/EntitySelector/styles/index.scss';
