+++ src/Campaigns/Blocks/shared/components/ModalForm/index.tsx
@@ -5,26 +5,23 @@ import {Button, Dialog, Modal, ModalOverlay} from 'react-aria-components';
 import ModalCloseIcon from './ModalClose';
 import {Spinner} from '@wordpress/components';
 import './styles.scss';
+import '../EntitySelector/styles/index.scss';
 
-import '../../../editor/styles/index.scss';
-
+/**
+ * @unreleasaed
+ */
 type ModalFormProps = {
     dataSrc: string;
     embedId: string;
-    openFormButton: string;
+    buttonText: string;
     isFormRedirect: boolean;
     formViewUrl: string;
 };
 
 /**
- * @since 4.1.0 updated with improved accessibility and styling
- * @since 4.0.0 updated to include loading state
- * @since 3.6.1
- * @since 3.4.0
- * @since 3.2.0 include types. update BEM classnames.
- * @since 3.0.0
+ * @unreleasaed
  */
-export default function ModalForm({dataSrc, embedId, openFormButton, isFormRedirect, formViewUrl}: ModalFormProps) {
+export default function ModalForm({dataSrc, embedId, buttonText, isFormRedirect, formViewUrl}: ModalFormProps) {
     const [dataSrcUrl, setDataSrcUrl] = useState(dataSrc);
     const [isOpen, setIsOpen] = useState<boolean>(isFormRedirect);
     const [isLoading, setLoading] = useState<boolean>(false);
@@ -104,7 +101,7 @@ export default function ModalForm({dataSrc, embedId, openFormButton, isFormRedir
                 )}
 
                 <span style={{margin: '0', visibility: isLoading ? 'hidden' : 'visible'}} aria-hidden={isLoading}>
-                    {openFormButton}
+                    {buttonText}
                 </span>
             </Button>
